requirements.txt
Flask==0.12.2
gunicorn==19.6.0
line-bot-sdk==1.3.0
matplotlib==2.0.2
numpy==1.13.1
pandas==0.20.3
psycopg2==2.6.2
requests==2.17.3
urllib3==1.21.1
Werkzeug==0.12.2
whitenoise==2.0.6
Pillow==4.2.1
image==1.5.11
